Add RemoteWebDriverProvider, and have SauceWebDriverProvider extend it to avoid duplicated code.
