+++ web-selenium/src/main/java/org/jbehave/web/selenium/SauceWebDriverProvider.java
@@ -1,49 +1,17 @@
 package org.jbehave.web.selenium;
 
-import org.openqa.selenium.Platform;
-import org.openqa.selenium.firefox.FirefoxDriver;
-import org.openqa.selenium.remote.CommandExecutor;
-import org.openqa.selenium.remote.DesiredCapabilities;
-import org.openqa.selenium.remote.HttpCommandExecutor;
-import org.openqa.selenium.remote.RemoteWebDriver;
-
 import java.net.MalformedURLException;
 import java.net.URL;
 
-public class SauceWebDriverProvider extends DelegatingWebDriverProvider {
-
-    DesiredCapabilities desiredCapabilities;
-
-    public SauceWebDriverProvider(DesiredCapabilities desiredCapabilities) {
-        this.desiredCapabilities = desiredCapabilities;
-    }
+public class SauceWebDriverProvider extends RemoteWebDriverProvider {
 
     public SauceWebDriverProvider() {
-        desiredCapabilities = DesiredCapabilities.firefox();
-        desiredCapabilities.setVersion("3.6.");
-        desiredCapabilities.setPlatform(Platform.WINDOWS);
+        super();
         desiredCapabilities.setCapability("name", "JBehave");
     }
 
-    public void initialize() {
-        try {
-            delegate.set(new RemoteWebDriver(createRemoteURL(), desiredCapabilities));
-        } catch (MalformedURLException e) {
-            banner();
-            e.printStackTrace();
-            throw new UnsupportedOperationException(e);
-        } catch (UnsupportedOperationException e) {
-            banner();
-            e.printStackTrace();
-            throw e;
-        } catch (Throwable e) {
-            banner();
-            e.printStackTrace();
-            throw new UnsupportedOperationException(e);
-        }
-    }
-
-    public static URL createRemoteURL() throws MalformedURLException {
+    @Override
+    public URL createRemoteURL() throws MalformedURLException {
         String username = System.getProperty("SAUCE_USERNAME");
         String access_key = System.getProperty("SAUCE_ACCESS_KEY");
         if (username == null) {
@@ -54,10 +22,6 @@ public class SauceWebDriverProvider extends DelegatingWebDriverProvider {
         }
 
         return new URL("http://" + username + ":" + access_key + "@ondemand.saucelabs.com/wd/hub");
-    }
-
-    private void banner() {
-        System.out.println("************** Error Initializing WebDriverProvider *************");
     }
 
 }
